+++ CHANGELOG.md
@@ -4,12 +4,18 @@ https://keepachangelog.com/en/1.0.0/
 
 ## 0.3.2-indev (????-??-??)
 
+### Added
+- Added flatpak support for Linux.
+
 ### Changed
 - The side panel is now resizable.
 - The side panel user interface is more responsive, and should be easier to read.
 
 ### Fixed
 - Fixed: The timer does not export correctly at export (#48).
+
+### Security
+- Updated Electron from v16.2.8 to v22.2.0
 
 ## 0.3.1-beta (2022-11-01)